Tweak popperOptions setting
